issues-55 fixed (#89)
* issues-55

A user can only revoke its own work request.

+++ src/main/java/com/kodekonveyor/work_request/revoke/RevokeWorkRequestController.java
@@ -7,6 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.kodekonveyor.authentication.AuthenticatedUserService;
+import com.kodekonveyor.authentication.UserEntity;
 import com.kodekonveyor.webapp.ValidationException;
 import com.kodekonveyor.work_request.WorkRequestConstants;
 import com.kodekonveyor.work_request.WorkRequestEntity;
@@ -18,13 +20,26 @@ public class RevokeWorkRequestController {
   @Autowired
   WorkRequestRepository workRequestRepository;
 
+  @Autowired
+  AuthenticatedUserService authenticatedUserService;
+
   @GetMapping("/workRequest/@workRequestId")
   public void call(@RequestParam final long workRequestId) {
     inputValidation(workRequestId);
 
     final WorkRequestEntity workRequestEntity =
         workRequestRepository.findByWorkRequestId(workRequestId).get(0);
-    workRequestRepository.delete(workRequestEntity);
+
+    final UserEntity user = authenticatedUserService.call();
+    final UserEntity customer = workRequestEntity.getCustomer();
+
+    if (customer.getId() == user.getId())
+      workRequestRepository.delete(workRequestEntity);
+    else
+      throw new ValidationException(
+          WorkRequestConstants.UNAUTHORIZE_REVOKE_WORK_REQUEST
+      );
+
   }
 
   private void inputValidation(final long workRequestId) {
